backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cohere>=5.20.1",
    "fastapi[standard]>=0.125.0",
    "lxml>=6.0.2",
    "openai-agents>=0.6.4",
    "python-dotenv>=1.2.1",
    "qdrant-client[fastembed]>=1.16.2",
    "playwright>=1.41.2",
    "requests>=2.32.5",
    "trafilatura>=2.0.0",
    "google-generativeai>=0.5.4",
    "uvicorn>=0.38.0",
]
